replace the override function with fb requests.
